add pandas into structure

+++ policyValueNet.py
@@ -93,8 +93,8 @@ class GenerateModelSeparateLastLayer:
 				valueLossSummary = tf.summary.scalar("valueLoss", valueLoss_)
 
 				relativeErrorBound_ = tf.constant(self.valueRelativeErrBound)
-				relativeValueError_ = tf.abs((valuePrediction_ - valueLabel_) / valueLabel_)
-				valueAccuracy_ = tf.reduce_mean(tf.cast(tf.less(relativeValueError_, relativeErrorBound_), tf.float32))
+				relativeValueError_ = tf.cast((tf.abs((valuePrediction_ - valueLabel_) / valueLabel_)), tf.float64)
+				valueAccuracy_ = tf.reduce_mean(tf.cast(tf.less(relativeValueError_, relativeErrorBound_), tf.float64))
 				tf.add_to_collection("valueAccuracy", valueAccuracy_)
 				valueAccuracySummary = tf.summary.scalar("valueAccuracy", valueAccuracy_)
